Added support for following other users.

+++ item/index.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 
 if($_SERVER['REQUEST_METHOD'] === 'DELETE') {
     doDelete();
@@ -15,30 +16,50 @@ function doDelete() {
         $keyspace = 'twitter';
         $session = $cluster->connect($keyspace);
         $statement = new Cassandra\SimpleStatement(
-            "DELETE FROM tweetsbyid WHERE id='" . $id . "' IF EXISTS"
+            "SELECT * FROM tweetsbyid WHERE id='" . $id . "'"
             );
         $future = $session->executeAsync($statement);
         $result = $future->get();
         $row = $result->first();
+        
+        if($row !== NULL) {
+            if(strcmp($_SESSION['username'], $row['username']) === 0) {
+               $statement = new Cassandra\SimpleStatement(
+                "DELETE FROM tweetsbyid WHERE id='" . $id . "' IF EXISTS"
+                );
+               $future = $session->executeAsync($statement);
+               $result = $future->get();
+               $row = $result->first();
 
-        if ($row['[applied]']) {
-            $phrase = 'OK';
+               if ($row['[applied]']) {
+                $phrase = 'OK';
+            }
+            else {
+                $phrase = 'ERROR';
+                $err = 'No tweet with id=' . $id;
+            }
         }
         else {
             $phrase = 'ERROR';
-            $err = 'No tweet with id=' . $id;
+            $err = 'You cannot delete tweets created by another user.';
         }
     }
     else {
-     $phrase = 'ERROR';
-     $err = 'No tweet id specified.';
- }
+        $phrase = 'ERROR';
+        $err = 'No tweet with id=' . $id;
+    }
 
- $session->closeAsync();
+}
+else {
+   $phrase = 'ERROR';
+   $err = 'No tweet id specified.';
+}
 
- $response = array("status" => $phrase);
+$session->closeAsync();
 
- if(strcmp($phrase, 'ERROR') === 0) {
+$response = array("status" => $phrase);
+
+if(strcmp($phrase, 'ERROR') === 0) {
     $response['error'] = $err;
 }
 $json = json_encode($response);
@@ -74,15 +95,15 @@ function doGet() {
         }
     }
     else {
-     $phrase = 'ERROR';
-     $err = 'No tweet id specified.';
- }
+       $phrase = 'ERROR';
+       $err = 'No tweet id specified.';
+   }
 
- $session->closeAsync();
+   $session->closeAsync();
 
- $response = array("status" => $phrase);
+   $response = array("status" => $phrase);
 
- if(strcmp($phrase, 'OK') === 0) {
+   if(strcmp($phrase, 'OK') === 0) {
     $response['item'] = array("id" => strval($id), "username" => $username, "content" => $content, "timestamp" => $timestamp);
 }
 else {
@@ -93,3 +114,4 @@ $json = json_encode($response);
 echo $json;
 }
 ?>
+
